requirements.txt
pywikibot
python-dotenv
requests
beautifulsoup4
pandas
